// sanity.types.ts
export type CateringPackage = {
  _id: string;
  _type: "cateringPackage";
  _createdAt: string;
  _updatedAt: string;
  _rev: string;
  name?: string;
  slug?: Slug;
  price?: number;
  priceUnit?: "porsi" | "orang";
  minOrder?: number;
  category?: {
    _ref: string;
    _type: "reference";
    _weak?: boolean;
    [internalGroqTypeReferenceTo]?: "category";
  };
  menuItems?: Array<{
    name?: string;
    description?: Array<string>;
    _key: string;
  }>;
  includes?: Array<string>;
  excludes?: Array<string>;
  isPopular?: boolean;
  isActive?: boolean;
  tags?: Array<string>;
  serviceArea?: Array<string>;
  advanceBooking?: number;
  featuredImage?: {
    asset?: {
      _ref: string;
      _type: "reference";
      _weak?: boolean;
      [internalGroqTypeReferenceTo]?: "sanity.imageAsset";
    };
    media?: unknown;
    hotspot?: SanityImageHotspot;
    crop?: SanityImageCrop;
    alt?: string;
    _type: "image";
  };
};
// ...
export type SanityImageHotspot = {
  _type: "sanity.imageHotspot";
  x?: number;
  y?: number;
  height?: number;
  width?: number;
};

export type SanityImageCrop = {
  _type: "sanity.imageCrop";
  top?: number;
  bottom?: number;
  left?: number;
  right?: number;
};
// ...
export type Slug = {
  _type: "slug";
  current?: string;
  source?: string;
};
// ...
export declare const internalGroqTypeReferenceTo: unique symbol;